Reduce startup time when running tests, making them much faster
This reduces the amount of time it takes to run tests. Before the
majority of the time was spent waiting for rails to boot inside the
container. But no longer!

This changes script/test to start a web container and keep it running in
the background and just ssh into that container when running tests. Then
rails is only loaded once and everything is good.

To stop or restart the rails container use script/kill-spring